Implementa camada de persistence
