+++ Invite.Persistence/Repositories/Interfaces/IPersonsRepository.cs
@@ -1,0 +1,7 @@
+using Invite.Entities.Models;
+
+namespace Invite.Persistence.Repositories.Interfaces;
+
+public interface IPersonsRepository : IGenericRepository<PersonModel>
+{
+}
