+++ Invite.Services/Invite.Services.csproj
@@ -14,6 +14,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Invite.Commons\Invite.Commons.csproj" />
     <ProjectReference Include="..\Invite.Entities\Invite.Entities.csproj" />
+    <ProjectReference Include="..\Invite.Persistence\Invite.Persistence.csproj" />
   </ItemGroup>
 
 </Project>
