--- api/Persistence/Repositories/Interfaces/IPersonsRepository.cs
@@ -1,7 +1,0 @@
-using api.Entities.Models;
-
-namespace api.Persistence.Repositories.Interfaces;
-
-public interface IPersonsRepository : IGenericRepository<PersonModel>
-{
-}
