--- api/Persistence/Repositories/Repository/PersonRepository.cs
@@ -1,12 +1,0 @@
-using api.Entities.Models;
-using api.Persistence.Context;
-using api.Persistence.Repositories.Interfaces;
-
-namespace api.Persistence.Repositories.Repository;
-
-public class PersonRepository(
-    AppDbContext context
-) : GenericRepository<PersonModel>(context),
-    IPersonsRepository
-{
-}
